fix: restore compilation across ~15 modules broken since PR #1356
The crate has not compiled since PR #1356 (staking/referral/auth-revocation,
merged 2026-07-30) — hidden because most subsequent PRs' changed paths never
triggered the path-filtered Contract Invariants workflow. This fixes the
~350 resulting compile errors:

- types.rs: add ~85 missing DataKey/ContractError variants that guarantor,
  staking_pool, vouch_syndication, flash_loan, bridge, multitoken_support,
  vouch_reputation, vouch_milestones, recurring_payment, referral,
  reputation_nft, prediction_market, community_treasury, governance_token,
  interest_rate_options, dynamic_interest, loyalty, and lib.rs's idempotency/
  rate-limit endpoints already referenced by name but were never declared.
  Also add the TokenBridgeMetadata/TokenSwapConfig structs and
  DEFAULT_LIQUIDITY_TIER_BONUSES const multitoken_support.rs needed.
- flash_loan.rs: fix calls to the single-arg token_client() helper, and
  swap an unauthorized transfer_from for a direct transfer for repayment.
- collateral_pool.rs, liquidity_mining.rs: require_admin_approval() returns
  (), not Result — drop the invalid `?`.
- circuit_breaker.rs, insurance.rs: fix `.unwrap_or(Ok(x))` on an Option
  (should just be `.unwrap_or(x)`), add missing Vec import, replace
  std::cmp::min (no_std crate) with the inherent .min().
- cross_chain_governance.rs: Vec<u8> isn't valid in #[contracttype] (u8 has
  no Val mapping) — use Bytes; add missing String import.
- arbitrage_prevention.rs: ExchangeRate derived Copy while holding Address
  fields, which isn't Copy — drop the derive.
- referral.rs: sha256() returns Hash<32>, not BytesN<32> — add .to_bytes().
- guarantor.rs: was looking up loans via the borrower-keyed
  get_active_loan_record(&Address) despite having a loan_id; added a
  get_loan_by_id() helper and used it instead.
- vouch.rs: Vec::get/set need u32 indices, Iterator::position() gives usize.
- lib.rs: add the 3 fields (default_rate_threshold, insurance_fund_premium_bps,
  insurance_max_payout_bps, all defaulted to 0/disabled) missing from the
  Config initializer; fix 6 stub endpoints constructing soroban_sdk::String
  via std's String::from("") instead of String::from_str(&env, ""); fix a
  moved-value use of `status` in batch loan status lookup; fix a u64/u32
  mismatch in an extend_ttl call.
- admin.rs, bridge.rs: symbol_short! literals over the 9-char limit.
- Remove audit_verification.rs (and its 3 lib.rs endpoints): it depends on
  crate::audit::get_vouch_audit_trail and VouchAuditEvent/VouchAuditEventType,
  none of which exist anywhere in the codebase — audit.rs is an explicit,
  documented stub ("implementation deferred pending definition of audit
  trail types"). Unlike the other modules here, there was no real logic
  to reconnect; building an audit-trail subsystem from scratch was out of
  scope for a compile fix.

+++ src/vouch.rs
@@ -2429,8 +2429,8 @@ pub fn get_portfolio_risk(env: Env, voucher: Address) -> PortfolioRiskReport {
 
         match token_addrs.iter().position(|t| t == v.token) {
             Some(i) => {
-                let updated = token_stakes.get(i).unwrap().saturating_add(v.stake);
-                token_stakes.set(i, updated);
+                let updated = token_stakes.get(i as u32).unwrap().saturating_add(v.stake);
+                token_stakes.set(i as u32, updated);
             }
             None => {
                 token_addrs.push_back(v.token.clone());
@@ -2440,8 +2440,8 @@ pub fn get_portfolio_risk(env: Env, voucher: Address) -> PortfolioRiskReport {
 
         match chain_ids.iter().position(|c| c == v.chain_id) {
             Some(i) => {
-                let updated = chain_stakes.get(i).unwrap().saturating_add(v.stake);
-                chain_stakes.set(i, updated);
+                let updated = chain_stakes.get(i as u32).unwrap().saturating_add(v.stake);
+                chain_stakes.set(i as u32, updated);
             }
             None => {
                 chain_ids.push_back(v.chain_id);
